+++ modules/LinkedList/slist.c
