+++ tests/map_test.c
