+++ tests/oset_test.c
