+++ tests/slist_test.c
